+++ mordant/src/commonMain/kotlin/com/github/ajalt/mordant/terminal/StdoutTerminalInterface.kt
@@ -6,8 +6,6 @@ import com.github.ajalt.mordant.rendering.AnsiLevel
 
 internal class StdoutTerminalInterface private constructor(
     override val info: TerminalInfo,
-    private val errInfo: TerminalInfo,
-    private val useStdErr: Boolean,
 ) : TerminalInterface {
     constructor(
         ansiLevel: AnsiLevel?,
@@ -15,30 +13,22 @@ internal class StdoutTerminalInterface private constructor(
         height: Int?,
         hyperlinks: Boolean?,
         interactive: Boolean?,
-    ) : this(TerminalDetection.detectTerminal(false, ansiLevel, width, height, hyperlinks, interactive),
-        TerminalDetection.detectTerminal(true, ansiLevel, width, height, hyperlinks, interactive), false)
-
-    constructor(): this(null, null, null, null, null)
+    ) : this(
+        TerminalDetection.detectTerminal(ansiLevel, width, height, hyperlinks, interactive)
+    )
 
     override fun completePrintRequest(request: PrintRequest) {
-        if (useStdErr) {
-            printStderr(request.text, request.trailingLinebreak)
-            return
-        }
-
-        if (request.trailingLinebreak) {
-            if (request.text.isEmpty()) {
-                println()
-            } else {
-                println(request.text)
+        when {
+            request.stderr -> printStderr(request.text, request.trailingLinebreak)
+            request.trailingLinebreak -> {
+                if (request.text.isEmpty()) {
+                    println()
+                } else {
+                    println(request.text)
+                }
             }
-        } else {
-            print(request.text)
+            else -> print(request.text)
         }
-    }
-
-    override fun forStdErr(): TerminalInterface {
-        return StdoutTerminalInterface(info, errInfo, true)
     }
 
     override fun readLineOrNull(hideInput: Boolean): String? = readLineOrNullMpp(hideInput)